feat: refactor registration flow by removing old components and implementing new RegisterForm with hooks
