--- src/features/login/hooks/login.ts
@@ -1,16 +1,0 @@
-import axiosInstance from "@/shared/lib/axios";
-
-export type SignUpResponseBody = {
-    name: string;
-    email: string;
-    password: string;
-}
-
-export const signUp = async (body: SignUpResponseBody) => {
-    try {
-        const response = await axiosInstance.post<SignUpResponseBody>("v1/interviewee", body);
-        return response.data;
-    } catch (error) {
-
-    }
-}
